[PATCH] Enable backlight toggling for Dell Inspiron 8200 GeForce2 Go
GeForce2 Go and similar cards are explicitly excluded from the
NVBacklightEnable() code (purposedly, as the code used for the other
chipsets on x86 has no effect.) This causes the backlight to remain
active on laptops equipped with such a card, even when DPMS modes were
active.

Fix the issue for Dell Inspiron 8200 by introducing the proper
registry writes to toggle the backlight.

Signed-off-by: Giuseppe Bilotta <giuseppe.bilotta@gmail.com>
Signed-off-by: Aaron Plattner <aplattner@nvidia.com>