The /compat code is no longer required.
This was used at one point in time to backport the modes code from server
1.3 to server 1.2. When the tarball was created, whatever modes code
was on the developer disk was included.

The server 1.2 as shipped in the tarball on the web does not contain the
modes code. It was added just after and found in git branch server-1.2-branch.

The modes code was initially included in version 2.0.95 and still compiles
with server 1.2. As of version 2.1.9 it no longer compiles.
Subsequent versions are not backward compatible and probably not meant to be.

The XRandR wiki points to articles saying that XRandR 1.2 first appeared in
server 1.3 and nv 2.0.95. Given that modes code in current server version
does not compile with server 1.2, and that the modes code is included in
version 1.3 and up, might as well stop including this modes code in the driver.

This patch makes the driver requiring server 1.3 or later.

This will solve the issue where "make dist" fails "out of the box"
and requires manual fix-up on the build machine which is documented
in the compat/README file.

Signed-off-by: Gaetan Nadon <memsize@videotron.ca>
Signed-off-by: Aaron Plattner <aplattner@nvidia.com>